// NotD-Aftermath-Web/src/app/models/codes/rank.ts
export enum Rank {
    None = 0,
    Pvt = 1,
    Pfc = 2,
    Lcpl = 3,
    Cpl = 4,
    Sgt = 5,
    SSgt = 6,
    GySgt = 7,
    Lt2 = 8,
    Lt1 = 9,
    Capt = 10,
    Maj = 11,
    LtCol = 12,
    Col = 13,
    BGen = 14,
    MajGen = 15,
    LtGen = 16,
    Gen = 17,
    FMar = 18,
    XpLimit = 19
}